Its definitly foolproof (sarcasm)
