+++ src/app/actions.ts
@@ -4,8 +4,7 @@ import { db } from "@/server/db";
 import type { UsersTableType } from "@/server/db/schema";
 import { users } from "@/server/db/schema";
 import type { PaperlessDocumentsType } from "@/types";
-import { auth, currentUser } from "@clerk/nextjs/server";
-import type { WhishperRecordingType } from "@/types";
+import { auth } from "@clerk/nextjs/server";
 
 /*
 Clerk helpers
@@ -78,38 +77,4 @@ export async function getPaperlessDocuments(query: string) {
   const data = (await response.json()) as PaperlessDocumentsType;
 
   return data;
-}
-
-export async function getWhishperRecordings(
-  query: string,
-): Promise<WhishperRecordingType[] | null> {
-  const userData = await getUserData();
-
-  if (!query || query == "null" || query.length < 3 || !userData) return null;
-
-  const response = await fetch(`${userData.whishperURL}/api/transcriptions`);
-
-  const data = (await response.json()) as WhishperRecordingType[];
-  const lowerCaseQuery = query.toLowerCase();
-  const filteredAndScored = data
-    .filter(
-      (item) =>
-        item.fileName.toLowerCase().includes(lowerCaseQuery) ||
-        item.result.text.toLowerCase().includes(lowerCaseQuery),
-    )
-    .map((item) => {
-      const fileNameOccurrences = (
-        item.fileName.toLowerCase().match(new RegExp(lowerCaseQuery, "g")) ?? []
-      ).length;
-      const textOccurrences = (
-        item.result.text.toLowerCase().match(new RegExp(lowerCaseQuery, "g")) ??
-        []
-      ).length;
-      const score = fileNameOccurrences + textOccurrences;
-      return { ...item, score };
-    });
-  const sortedByScore = filteredAndScored.sort((a, b) => b.score - a.score);
-
-  // Step 4: Return the sorted array without the score
-  return sortedByScore.map(({ ...item }) => item);
 }
