+++ src/app/whishper/page.tsx
@@ -22,7 +22,7 @@ import {
   QueryClientProvider,
   QueryClient,
 } from "@tanstack/react-query";
-import { getUserData, getWhishperRecordings } from "../actions";
+import { getUserData } from "../actions";
 import LoadingSpinner from "@/components/loading-spinner";
 import OpenInternalLink from "@/components/internal-link";
 import {
@@ -45,6 +45,40 @@ import { BadgeCheck, Badge, BadgeAlert } from "lucide-react";
 import type { WhishperRecordingType } from "@/types";
 
 const queryClient = new QueryClient();
+
+export async function getWhishperRecordings(
+  query: string,
+): Promise<WhishperRecordingType[] | null> {
+  const userData = await getUserData();
+
+  if (!query || query == "null" || query.length < 3 || !userData) return null;
+
+  const response = await fetch(`${userData.whishperURL}/api/transcriptions`);
+
+  const data = (await response.json()) as WhishperRecordingType[];
+  const lowerCaseQuery = query.toLowerCase();
+  const filteredAndScored = data
+    .filter(
+      (item) =>
+        item.fileName.toLowerCase().includes(lowerCaseQuery) ||
+        item.result.text.toLowerCase().includes(lowerCaseQuery),
+    )
+    .map((item) => {
+      const fileNameOccurrences = (
+        item.fileName.toLowerCase().match(new RegExp(lowerCaseQuery, "g")) ?? []
+      ).length;
+      const textOccurrences = (
+        item.result.text.toLowerCase().match(new RegExp(lowerCaseQuery, "g")) ??
+        []
+      ).length;
+      const score = fileNameOccurrences + textOccurrences;
+      return { ...item, score };
+    });
+  const sortedByScore = filteredAndScored.sort((a, b) => b.score - a.score);
+
+  // Step 4: Return the sorted array without the score
+  return sortedByScore.map(({ ...item }) => item);
+}
 
 function SearchForm() {
   const formSchema = z.object({
