+++ src/components/audio-preview.tsx
@@ -9,6 +9,13 @@ import type { UsersTableType } from "@/server/db/schema";
 import { Button } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
 import { ChevronLeft, ExternalLink } from "lucide-react";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "@/components/ui/tooltip";
+import { useState } from "react";
 
 const queryClient = new QueryClient();
 
@@ -48,6 +55,7 @@ function SkeletonLoader() {
 
 function AudioInfo(props: { name: string }) {
   const router = useRouter();
+  const [isPlaying, setIsPlaying] = useState(false);
 
   const {
     data: userData,
@@ -74,7 +82,9 @@ function AudioInfo(props: { name: string }) {
       <div className="flex h-full justify-center md:w-1/2">
         <div className="flex h-full w-full flex-col rounded-xl bg-slate-600/50">
           <div className="m-4 flex h-full flex-grow flex-col justify-center gap-9">
-            <div className="flex flex-row items-center justify-between gap-2 rounded-md bg-gradient-to-r from-indigo-900 to-purple-900 px-2 py-1 text-xl text-white">
+            <div
+              className={`audioPreview flex flex-row items-center justify-between gap-2 rounded-md px-2 py-1 text-xl text-white ${isPlaying ? "bg-gradient-to-r from-indigo-900 to-purple-900" : ""}`}
+            >
               {formattedName}
               <Button
                 className="rounded-full border-none bg-slate-500/60 hover:bg-slate-500/90"
@@ -87,7 +97,12 @@ function AudioInfo(props: { name: string }) {
             </div>
             {/* Audio */}
             <div className="flex w-full flex-grow justify-center">
-              <audio controls={true} className="w-full md:w-3/4">
+              <audio
+                controls={true}
+                className="w-full md:w-3/4"
+                onPlay={() => setIsPlaying(true)}
+                onPause={() => setIsPlaying(false)}
+              >
                 <source
                   src={`${userData.whishperURL}/api/video/${props.name}`}
                   type="audio/mp4"
@@ -95,13 +110,46 @@ function AudioInfo(props: { name: string }) {
               </audio>
             </div>
             <div className="flex w-full flex-shrink-0 justify-center gap-12">
-              <Button className="w-24">
-                <ExternalLink /> Open
-              </Button>
-              <Button className="w-24">Download</Button>
-              <Button className="w-24" variant="destructive">
-                Delete
-              </Button>
+              <TooltipProvider delayDuration={50}>
+                <Tooltip>
+                  <TooltipTrigger
+                    aria-disabled="true"
+                    tabIndex={-1}
+                    className="cursor-not-allowed opacity-50"
+                  >
+                    <Button className="w-24">
+                      <ExternalLink /> Open
+                    </Button>
+                  </TooltipTrigger>
+                  <TooltipContent>Comming soon!</TooltipContent>
+                </Tooltip>
+              </TooltipProvider>
+              <TooltipProvider delayDuration={50}>
+                <Tooltip>
+                  <TooltipTrigger
+                    aria-disabled="true"
+                    tabIndex={-1}
+                    className="cursor-not-allowed opacity-50"
+                  >
+                    <Button className="w-24">Download</Button>
+                  </TooltipTrigger>
+                  <TooltipContent>Comming soon!</TooltipContent>
+                </Tooltip>
+              </TooltipProvider>
+              <TooltipProvider delayDuration={50}>
+                <Tooltip>
+                  <TooltipTrigger
+                    aria-disabled="true"
+                    tabIndex={-1}
+                    className="cursor-not-allowed opacity-50"
+                  >
+                    <Button className="w-24" variant="destructive">
+                      Delete
+                    </Button>
+                  </TooltipTrigger>
+                  <TooltipContent>Comming soon!</TooltipContent>
+                </Tooltip>
+              </TooltipProvider>
             </div>
           </div>
         </div>
